+++ src/main/java/com/biomatiques/controller/ShiftController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.biomatiques.model.Login;
 import com.biomatiques.model.Shift;
 import com.biomatiques.services.ShiftService;
 
@@ -24,35 +25,65 @@ public class ShiftController {
 	    
 	    @RequestMapping(value= {"/shift.html"},method=RequestMethod.GET)
 		public String shiftHomePage() {
-			return "shift";
+	    	if(Login.loggedin==true) {
+	    		return "shift";
+				}
+				else {
+					return "error1.html";
+				}
+			
 		}
 	    
 	    @RequestMapping(value="/shiftForm.html",method=RequestMethod.GET)
 	    public String shiftForm(Shift shift) {
-	    	return "shiftForm";
+	    	if(Login.loggedin==true) {
+	    		return "shiftForm";
+				}
+				else {
+					return "error1.html";
+				}
+			
+	    	
 	    }
 	  //Add Method
 	    @RequestMapping(value="/addShift",method=RequestMethod.POST)
 	    public String addShift( @Valid Shift shift,BindingResult result,Model model) {
 	    	/*if(result.hasErrors()) {
 	    		return "add-shift";
-	    	}*/
-	        shiftService.addShift(shift);
-	        model.addAttribute("shift", shiftService.getAllShifts());
-	        return "redirect:/viewShift.html";
+	    	}*/	       
+	        if(Login.loggedin==true) {
+	        	 shiftService.addShift(shift);
+	 	        model.addAttribute("shift", shiftService.getAllShifts());
+	 	        return "redirect:/viewShift.html";
+				}
+				else {
+					return "error1.html";
+				}
 	    }
 	    //GET Method
 	    @RequestMapping(method = RequestMethod.GET, value = "/viewShift.html")
 	    public String getAllEmployees(Model model){
-	        model.addAttribute("shifts", shiftService.getAllShifts());
-	        return "viewShift";
+	    	 if(Login.loggedin==true) {
+	    		 model.addAttribute("shifts", shiftService.getAllShifts());
+	 	        return "viewShift";
+				}
+				else {
+					return "error1.html";
+				}
+	       
 	    }
 	    
 	    //Edit Form
 	    @RequestMapping(value="/editShift/{id}",method=RequestMethod.GET)
 	    public String editShiftById(@PathVariable Long id,Model model) {
-	    	model.addAttribute("shift",shiftService.getShiftById(id));
-	    	return "editShift";
+	    	if(Login.loggedin==true) {
+	    		model.addAttribute("shift",shiftService.getShiftById(id));
+		    	return "editShift";
+				}
+				else {
+					return "error1.html";
+				}
+	    	
 	    }
 	    
 	    
@@ -62,25 +93,43 @@ public class ShiftController {
 	    	/*if(result.hasErrors()) {
 	    		return "update-shift";
 	    	}*/
-	        shiftService.updateShift(shift);
-	        model.addAttribute("shift", shiftService.getAllShifts());
-	        return "redirect:/viewShift.html";
+	    	if(Login.loggedin==true) {
+	    		 shiftService.updateShift(shift);
+	 	        model.addAttribute("shift", shiftService.getAllShifts());
+	 	        return "redirect:/viewShift.html";
+				}
+				else {
+					return "error1.html";
+				}
+	       
 	    }
 	    
 	    
 	    //GET SHIFT BY ID
 	    @RequestMapping(value="/shift/{id}",method=RequestMethod.GET)
 	    public String getShiftById(@PathVariable Long id,Model model) {
-	    	model.addAttribute("shift",shiftService.getShiftById(id));
-	    	return "updateShift";
+	    	if(Login.loggedin==true) {
+	    		model.addAttribute("shift",shiftService.getShiftById(id));
+		    	return "updateShift";
+				}
+				else {
+					return "error1.html";
+				}
+	    	
 	    }
 	  
 	    
 	    //Delete Method
 	    @RequestMapping(value="/deleteShift/{id}",method=RequestMethod.GET)
 	    public String deleteEmployee( @PathVariable long id,Model model) {
-	        shiftService.deleteShift(id);
-	        model.addAttribute("shifts", shiftService.getAllShifts());
-	        return "redirect:/viewShift.html";
+	    	if(Login.loggedin==true) {
+	    		 shiftService.deleteShift(id);
+	 	        model.addAttribute("shifts", shiftService.getAllShifts());
+	 	        return "redirect:/viewShift.html";
+				}
+				else {
+					return "error1.html";
+				}
+	       
 	    }
 }
