+++ src/main/java/com/biomatiques/controller/ShiftPatternController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import com.biomatiques.model.Login;
 import com.biomatiques.model.ShiftPattern;
 import com.biomatiques.services.ShiftPatternService;
 
@@ -26,39 +27,69 @@ public class ShiftPatternController {
 	
 	@RequestMapping(method = RequestMethod.GET, value = "/shiftPatternForm.html")
     public String shiftPatternForm(ShiftPattern shiftPattern){
-        return "shiftPatternForm";
+		if(Login.loggedin==true) {
+			 return "shiftPatternForm";
+			}
+			else {
+				return "error1.html";
+			}
+       
     }
 	
 	//Add Method
     @RequestMapping(value="/addShiftPattern",method=RequestMethod.POST)
     public String addShiftPattern( @Valid ShiftPattern shiftPattern ,BindingResult result,Model model) {
-        if(result.hasErrors()) {
-        	return "viewShiftPattern";
-        }
-        shiftPatternService.addShiftPattern(shiftPattern);
-        model.addAttribute("shiftPattern", shiftPatternService.getAllShiftPattern());
-        return "redirect:/viewShiftPattern.html";    	
+    	if(Login.loggedin==true) {
+    		 if(result.hasErrors()) {
+    	        	return "viewShiftPattern";
+    	        }
+    	        shiftPatternService.addShiftPattern(shiftPattern);
+    	        model.addAttribute("shiftPattern", shiftPatternService.getAllShiftPattern());
+    	        return "redirect:/viewShiftPattern.html";  
+			}
+			else {
+				return "error1.html";
+			}
+         	
     }
     
   //Edit Form
     @RequestMapping(value="/editShiftPattern/{id}",method=RequestMethod.GET)
     public String editShiftById(@PathVariable Long id,Model model) {
-    	model.addAttribute("shiftPattern",shiftPatternService.getShiftPatternById(id));
-    	return "editShiftPattern";
+    	if(Login.loggedin==true) {
+    		model.addAttribute("shiftPattern",shiftPatternService.getShiftPatternById(id));
+        	return "editShiftPattern";
+			}
+			else {
+				return "error1.html";
+			}
+    	
     }
     
     //GET Methods one and all employees
     @RequestMapping(method = RequestMethod.GET, value = "/viewShiftPattern.html")
     public String getAllShiftPattern(Model model){
-        model.addAttribute("shiftPattern",shiftPatternService.getAllShiftPattern());
-        return "viewShiftPattern";
+    	if(Login.loggedin==true) {
+    		 model.addAttribute("shiftPattern",shiftPatternService.getAllShiftPattern());
+    	        return "viewShiftPattern";
+			}
+			else {
+				return "error1.html";
+			}
+       
     }
     
     
     @RequestMapping(value="/shiftPattern/{id}",method=RequestMethod.GET)
     public String getShiftPatternById(@PathVariable Long id,Model model) {
-    	model.addAttribute("shiftPattern",shiftPatternService.getShiftPatternById(id));
-        return "view-shiftPattern";
+    	if(Login.loggedin==true) {
+    		model.addAttribute("shiftPattern",shiftPatternService.getShiftPatternById(id));
+            return "view-shiftPattern";
+			}
+			else {
+				return "error1.html";
+			}
+    
     }  
     
     
@@ -68,16 +99,28 @@ public class ShiftPatternController {
 //    	if(result.hasErrors()) {
 //        	return "shiftPattern-home";
 //        }
-    	shiftPatternService.updateShiftPattern(shiftPattern);
-    	 model.addAttribute("shiftPattern",shiftPatternService.getAllShiftPattern());
-         return "redirect:/viewShiftPattern.html";
+    	if(Login.loggedin==true) {
+    		shiftPatternService.updateShiftPattern(shiftPattern);
+    		model.addAttribute("shiftPattern",shiftPatternService.getAllShiftPattern());
+            return "redirect:/viewShiftPattern.html";
+			}
+			else {
+				return "error1.html";
+			}
+    	
     }
     
     //Delete Method
     @RequestMapping(value="/deleteShiftPattern/{id}",method=RequestMethod.GET)
     public String deleteShiftPattern( @PathVariable long id,Model model) {
-        shiftPatternService.deleteShiftPattern(id);
-        model.addAttribute("shiftPattern",shiftPatternService.getAllShiftPattern());
-        return "redirect:/viewShiftPattern.html";
+    	if(Login.loggedin==true) {
+    		 shiftPatternService.deleteShiftPattern(id);
+    	        model.addAttribute("shiftPattern",shiftPatternService.getAllShiftPattern());
+    	        return "redirect:/viewShiftPattern.html";
+			}
+			else {
+				return "error1.html";
+			}
+       
     }
 }
